Fix GLcontext lost when app is put in background
In order to restore the GL context the "SDL_GL_CreateContextGL" function
is called after a "nativeResume". The "createEGLContext" function is
then called on the java side.

comments from SDL_androidevents.c:
createEGLContext will attempt to restore the GL context first, and if
that fails it will create a new one and if that fails it will create a
new one.
If a new GL context is created, the user needs to restore the textures
manually.

I remove the SDL_ANDROID_BLOCK_ON_PAUSE define as we already block the
core via "emuPause".